@@ -9,12 +9,12 @@ import (
 )
 
 type notBlankTest struct {
-	String    string      `validate:"notblank"`
-	Array     []int       `validate:"notblank"`
-	Pointer   *int        `validate:"notblank"`
-	Number    int         `validate:"notblank"`
-	Interface interface{} `validate:"notblank"`
-	Func      func()      `validate:"notblank"`
+	String    string `validate:"notblank"`
+	Array     []int  `validate:"notblank"`
+	Pointer   *int   `validate:"notblank"`
+	Number    int    `validate:"notblank"`
+	Interface any    `validate:"notblank"`
+	Func      func() `validate:"notblank"`
 }
 
 func TestNotBlank(t *testing.T) {
